[irssi-conf] Autojoin in config, not startup
Before, the startup file tries connecting to the default network and
join two channels. However, since irssi just executes the startup file
line by line with little logic behind it, it tries joining the channels
while the user is still connecting.

Instead, this commit sets the channels to autojoin, which only triggers
once the user is connected and, if configured properly, authenticated.